requirements.txt
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.4
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
